feat: Stat Values in DumpedItem
- stat are filled with values inside the dumped item

+++ src/main/java/poeai/stat/files/Stat.java
@@ -1,6 +1,6 @@
 package poeai.stat.files;
 
 public record Stat(String id,
-                   Integer value) {
+                   Double value) {
 
 }
